refactor: rename to path-builders.ts
